requirements.txt
flask
python-telegram-bot
python-dotenv
pandas
openpyxl
pystray
Pillow
